Add plumbing for wms endpoint and layers
Added props to the various deck gl maps across the plugin. Updated the
controlPanel files to remove mapbox styles field since it will not be
used and added a wms endpoint string field and wms layers multiselect.
Updated the store to provide an empty array for default wms layers
value.