Add wms results to redux
Adding geoserver wms fetch to redux to simplify the plumbing of wms
layer options. These options are cached so it's not recalled every time
a user need to get the layer options for a new chart.

Adding xml2js in order to parse the geoserver response. I also had to
add timers manually since it's required by xml2js but no listed as a
dependency. :facepalm